model constructors add id param

+++ server/src/models/user.ts
@@ -1,20 +1,42 @@
 import DataModel from './data.js';
+import type { PoolClient } from 'pg';
 import { StatusError } from '../util/error.js';
 import { Profile, User, MinUser } from 'oi-types/user';
 import { ContactKind } from 'oi-types/user';
+import { EventSummary } from 'oi-types/event';
 import bcrypt from 'bcrypt';
-import crypto from 'node:crypto';
 
 export default class UserModel extends DataModel {
-  async getProfileForSession(session: string): Promise<Profile> {
+  private _uid: string;
+
+  constructor(client: PoolClient, uid: string) {
+    super(client);
+    this._uid = uid;
+  }
+
+  async getEventSummary(): Promise<EventSummary[]> {
+    const result = await this._dbclient.query<EventSummary>(
+      `
+      SELECT events.id, events.title, events.start_time AS "startTime",
+        events.end_time AS "endTime", events.public, attendance.kind
+      FROM attendance
+      INNER JOIN events ON attendance.eid = events.id
+      WHERE attendance.uid = $1 AND events.gid IS NULL
+      `,
+      [this._uid]
+    );
+    return result.rows;
+  }
+
+  async getProfile(): Promise<Profile> {
     const result = await this._dbclient.query<Profile>(
       `
       SELECT id, name, username, avatar_url AS "avatarUrl", email, phone,
         verified, public
       FROM users
-      WHERE id = (SELECT uid FROM sessions WHERE sesskey = $1)
+      WHERE id = $1
       `,
-      [session]
+      [this._uid]
     );
     if (!result.rowCount) {
       throw new StatusError(404, 'User not found');
@@ -22,17 +44,17 @@ export default class UserModel extends DataModel {
     return result.rows[0];
   }
 
-  async getContactForSession(session: string, contactUid: string): Promise<User> {
+  async getContact(contactUid: string): Promise<User> {
     const result = await this._dbclient.query<User>(
       `
       SELECT users.id, users.name, users.username, users.avatar_url AS "avatarUrl",
         contacts.kind
       FROM users
       LEFT JOIN contacts ON users.id = contacts.uid_contact
-        AND contacts.uid_owner = (SELECT uid FROM sessions WHERE sesskey = $2)
+        AND contacts.uid_owner = $2
       WHERE users.id = $1
       `,
-      [contactUid, session]
+      [contactUid, this._uid]
     );
     if (result.rowCount === 0) {
       throw new StatusError(404, 'User not found');
@@ -40,7 +62,7 @@ export default class UserModel extends DataModel {
     return result.rows[0];
   }
 
-  async getContactsForUser(uid: string): Promise<User[]> {
+  async getContacts(): Promise<User[]> {
     const result = await this._dbclient.query<User>(
       `
       SELECT users.id, users.name, users.username,
@@ -49,12 +71,12 @@ export default class UserModel extends DataModel {
       INNER JOIN users ON contacts.uid_contact = users.id
       WHERE contacts.uid_owner = $1
       `,
-      [uid]
+      [this._uid]
     );
     return result.rows;
   }
 
-  async getFollowersForUser(uid: string): Promise<User[]> {
+  async getFollowers(): Promise<User[]> {
     const result = await this._dbclient.query<User>(
       `
       SELECT users.id, users.name, users.username,
@@ -63,33 +85,33 @@ export default class UserModel extends DataModel {
       INNER JOIN users ON contacts.uid_owner = users.id
       WHERE contacts.uid_contact = $1
       `,
-      [uid]
+      [this._uid]
     );
     return result.rows;
   }
 
-  async getContactRequestsForSession(session: string): Promise<User[]> {
+  async getContactRequests(): Promise<User[]> {
     const result = await this._dbclient.query(
       `
       SELECT users.id, users.name, users.username,
         users.avatar_url AS "avatarUrl"
       FROM contacts
       INNER JOIN users ON contacts.uid_owner = users.id
-      WHERE contacts.uid_contact = (SELECT uid FROM sessions WHERE sesskey = $1)
+      WHERE contacts.uid_contact = $1
         AND contacts.kind = 0
       `,
-      [session]
+      [this._uid]
     );
     return result.rows;
   }
 
-  async sessionHasContact(session: string, uidContact: string): Promise<boolean> {
+  async hasContact(uidContact: string): Promise<boolean> {
     const result = await this._dbclient.query(
       `
       SELECT kind FROM contacts
-      WHERE (uid_owner, uid_contact) = ((SELECT uid FROM sessions WHERE sesskey = $1), $2)
+      WHERE (uid_owner, uid_contact) = ($1, $2)
       `,
-      [session, uidContact]
+      [this._uid, uidContact]
     );
     if (result.rowCount === 0 || result.rows[0].kind === ContactKind.Requested) {
       return false;
@@ -97,10 +119,10 @@ export default class UserModel extends DataModel {
     return true;
   }
 
-  async getMinUser(uid: string): Promise<MinUser> {
+  async getMinUser(): Promise<MinUser> {
     const result = await this._dbclient.query<MinUser>(
       `SELECT id, name, username, avatar_url AS "avatarUrl" FROM users WHERE id = $1`,
-      [uid]
+      [this._uid]
     );
     if (result.rowCount === 0) {
       throw new StatusError(404, 'User not found');
@@ -108,10 +130,10 @@ export default class UserModel extends DataModel {
     return result.rows[0];
   }
 
-  async changePassword(uid: string, oldPassword: string, newPassword: string) {
+  async changePassword(oldPassword: string, newPassword: string) {
     const pwhashResult = await this._dbclient.query<{ pwhash: string | null }>(
       `SELECT pwhash FROM users WHERE id = $1`,
-      [uid]
+      [this._uid]
     );
     if (!pwhashResult.rowCount) {
       throw new StatusError(404, 'User not found');
@@ -127,22 +149,22 @@ export default class UserModel extends DataModel {
     const pwhash = await bcrypt.hash(newPassword, 10);
     const updateResult = await this._dbclient.query(
       `UPDATE users SET pwhash = $1 WHERE id = $2`,
-      [pwhash, uid]
+      [pwhash, this._uid]
     );
     if (updateResult.rowCount === 0) {
       throw new StatusError(500, 'Failed to update password');
     }
   }
 
-  async updateUserField(uid: string, field: string, value: any) {
+  async updateUserField(field: string, value: any) {
     await this._dbclient.query(
       `UPDATE users SET "${field}" = $1 WHERE id = $2`,
-      [value, uid]
+      [value, this._uid]
     );
   }
 
-  async getIdForUsername(username: string): Promise<string | null> {
-    const result = await this._dbclient.query<{ id: string }>(
+  static async getIdForUsername(dbclient: PoolClient, username: string): Promise<string | null> {
+    const result = await dbclient.query<{ id: string }>(
       `SELECT id FROM users WHERE lower(username) = lower($1)`,
       [username]
     )
@@ -152,16 +174,14 @@ export default class UserModel extends DataModel {
     return result.rows[0].id;
   }
 
-  async getIdAndAvatarForSession(session: string):
-      Promise<{ id: string; avatar_url: string | null; } | null> {
+  async getIdAndAvatar(): Promise<{ id: string; avatar_url: string | null; } | null> {
     const result = await this._dbclient.query<{ id: string; avatar_url: string | null; }>(
       `
       SELECT id, avatar_url
       FROM users
-      INNER JOIN sessions ON users.id = sessions.uid
-      WHERE sesskey = $1
+      WHERE id = $1
       `,
-      [session]
+      [this._uid]
     );
     if (!result.rowCount) {
       return null;
@@ -169,9 +189,9 @@ export default class UserModel extends DataModel {
     return result.rows[0];
   }
 
-  async getNameAndPublic(uid: string): Promise<{ name: string; public: string }> {
+  async getNameAndPublic(): Promise<{ name: string; public: string }> {
     const result = await this._dbclient.query<{ name: string; public: string }>(
-      `SELECT name, public FROM users WHERE id = $1`, [uid]
+      `SELECT name, public FROM users WHERE id = $1`, [this._uid]
     );
     if (!result.rowCount) {
       throw new StatusError(404, 'User not found');
@@ -179,9 +199,9 @@ export default class UserModel extends DataModel {
     return result.rows[0];
   }
 
-  async getUsername(uid: string): Promise<string> {
+  async getUsername(): Promise<string> {
     const result = await this._dbclient.query<{ username: string }>(
-      `SELECT username FROM users WHERE id = $1`, [uid]
+      `SELECT username FROM users WHERE id = $1`, [this._uid]
     );
     if (!result.rowCount) {
       throw new StatusError(404, 'User not found');
@@ -189,9 +209,9 @@ export default class UserModel extends DataModel {
     return result.rows[0].username;
   }
 
-  async getNameAndUsername(uid: string): Promise<{ name: string; username: string }> {
+  async getNameAndUsername(): Promise<{ name: string; username: string }> {
     const result = await this._dbclient.query<{ name: string; username: string }>(
-      `SELECT name, username FROM users WHERE id = $1`, [uid]
+      `SELECT name, username FROM users WHERE id = $1`, [this._uid]
     );
     if (!result.rowCount) {
       throw new StatusError(404, 'User not found');
@@ -199,45 +219,39 @@ export default class UserModel extends DataModel {
     return result.rows[0];
   }
 
-  async addContact(myUid: string, contactUid: string, kind: ContactKind) {
+  async addContact(contactUid: string, kind: ContactKind) {
     const result = await this._dbclient.query(
       `
       INSERT INTO contacts (uid_owner, uid_contact, kind)
       VALUES ($1, $2, $3)
       `,
-      [myUid, contactUid, kind]
+      [this._uid, contactUid, kind]
     );
     if (!result.rowCount) {
       throw new StatusError(400, 'Contacts failed to update');
     }
   }
 
-  async deleteContact(session: string, contactUid: string) {
+  async deleteContact(contactUid: string) {
     const result = await this._dbclient.query(
-      `
-      DELETE FROM contacts
-      WHERE uid_owner = (SELECT uid FROM sessions WHERE sesskey = $1)
-        AND uid_contact = $2
-      `,
-      [session, contactUid]
+      `DELETE FROM contacts WHERE uid_owner = $1 AND uid_contact = $2`,
+      [this._uid, contactUid]
     );
     if (result.rowCount === 0) {
       throw new StatusError(404, 'Contact not found');
     }
   }
 
-  async approveContact(session: string, uidOwner: string):
-      Promise<{ uid_contact: string; name: string}> {
+  async approveContact(uidOwner: string): Promise<{ uid_contact: string; name: string}> {
     const result = await this._dbclient.query<{ uid_contact: string; name: string }>(
       `
       UPDATE contacts SET contacts.kind = 1
       FROM users
-      WHERE (contacts.uid_owner, contacts.uid_contact) =
-        ($1, (SELECT uid FROM sessions WHERE sesskey = $2))
+      WHERE (contacts.uid_owner, contacts.uid_contact) = ($1, $2)
         AND users.id = contacts.uid_contact
       RETURNING contacts.uid_contact, users.name
       `,
-      [uidOwner, session]
+      [uidOwner, this._uid]
     );
     if (!result.rowCount) {
       throw new StatusError(404, 'Contact not found');
@@ -245,32 +259,20 @@ export default class UserModel extends DataModel {
     return result.rows[0];
   }
 
-  async denyContact(session: string, uidOwner: string):
-      Promise<{ uid_contact: string; name: string }> {
+  async denyContact(uidOwner: string): Promise<{ uid_contact: string; name: string }> {
     const result = await this._dbclient.query<{ uid_contact: string; name: string }>(
       `
       DELETE FROM contacts
       USING users
-      WHERE (uid_owner, uid_contact) = ($1, (SELECT uid FROM sessions WHERE sesskey = $2))
+      WHERE (uid_owner, uid_contact) = ($1, $2)
         AND users.id = uid_contact
       RETURNING uid_contact, users.name
       `,
-      [uidOwner, session]
+      [uidOwner, this._uid]
     );
     if (result.rowCount === 0) {
       throw new StatusError(404, 'Contact not found');
     }
     return result.rows[0];
-  }
-
-  static async createAvatarFilename(): Promise<string> {
-    const promise = new Promise<string>((resolve, reject) => {
-      crypto.randomBytes(32, (err, buf) => {
-        if (err) reject(err);
-        const name = buf.toString('base64').replaceAll(/\//g, '_').replaceAll(/\+/g, '-');
-        resolve(name);
-      });
-    });
-    return promise;
   }
 }
